backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pdfplumber
python-docx
scikit-learn
aiofiles
requests
python-dotenv
